Update README.md
## Part 1: Imaging Technology Inspiration

I chose the scene from the film *Inception* where the dreams are layered. I wanted to introduce this layering and visual depth into my project to enhance the user experience. This technique can be achieved by using clever animations, parallax scrolling, and perspective effects to create a more immersive experience for the user.

![Inception Dream Layer](https://media.licdn.com/dms/image/C4E12AQG0XLL_RbB5WA/article-cover_image-shrink_720_1280/0/1528191623683?e=2147483647&v=beta&t=iqOMnKhi5j5wOa5YsrDSEqrUqv0HI8mjVtpw8Y2skpc)
![Inception Dream Layer](https://www.escapistmagazine.com/wp-content/uploads/2020/07/inception1.jpg?fit=1200%2C675)

## Part 2: Exploring Coding Techniques

I'll explore achieving this sense of layering and visual depth using Three.js, a JavaScript library for creating 3D graphics that makes it easy to create complex scenes and animation effects. By taking advantage of the camera and scene features in Three.js, we can achieve a variety of layering and visual effects.

[Three.js Example](https://threejs.org/examples/)